refactor: update type & data source
