+++ plugin/web/extensions/sidebar/core/components/content/common/DatasetCard/index.tsx
@@ -75,7 +75,7 @@ const DatasetCard: React.FC<Props> = ({
         icon: "search",
         value: 1,
         onClick: () => {
-          onThreeDTilesSearch(dataset.dataId);
+          onThreeDTilesSearch(dataset.dataID);
         },
       });
     }
